do some things that make sense to me
