+++ src/Navigator/NavigationContainer/index.js
@@ -1,0 +1,3 @@
+import RouterContainer from './RouterContainer';
+
+export default RouterContainer;
